nharker-core: moved exceptions to interfaces
