+++ nharker-core/src/main/kotlin/com/tsbonev/nharker/adapter/nitrite/NitriteArticles.kt
@@ -1,7 +1,6 @@
 package com.tsbonev.nharker.adapter.nitrite
 
 import com.tsbonev.nharker.core.*
-import com.tsbonev.nharker.core.exceptions.*
 import com.tsbonev.nharker.core.helpers.*
 import org.dizitart.kno2.filters.eq
 import org.dizitart.no2.Nitrite
