+++ nharker-core/src/main/kotlin/com/tsbonev/nharker/core/Catalogues.kt
@@ -1,6 +1,5 @@
 package com.tsbonev.nharker.core
 
-import com.tsbonev.nharker.core.exceptions.*
 import java.util.Optional
 
 /**
