+++ nharker-core/src/main/kotlin/com/tsbonev/nharker/core/Entry.kt
@@ -20,3 +20,7 @@ data class Entry (@Id val id: String,
                   val articleId: String = ReferenceId.None.value,
                   val content: String,
                   val links: Map<String, String>)
+
+class EntryAlreadyInArticleException : Exception()
+class EntryNotFoundException : Exception()
+class EntryNotInArticleException : Exception()
