--- nharker-core/src/main/kotlin/com/tsbonev/nharker/core/exceptions/ArticleNotFoundException.kt
@@ -1,6 +1,0 @@
-package com.tsbonev.nharker.core.exceptions
-
-/**
- * @author Tsvetozar Bonev (tsbonev@gmail.com)
- */
-class ArticleNotFoundException : Exception()
